Move redux jump location into its own cache
